requirements.txt
# This file may be used to pip install dependencies using:
# $ pip install -r requirements.txt

colorama
cython
h5py
humanize
jinja2
jupyter
matplotlib
# mpi4py
numpy<2
numpy-stl
pandas
pre-commit
psutil
# pycuda
# pyopencl
scipy
terminaltables
tqdm
wheel
reframe-hpc
pytest
pytest-cov
# pytest-benchmark
# pytest-benchmark[histogram]
# pytest-regressions
git+https://github.com/craig-warren/PyEVTK.git
